Move constants to constants.js in poly-loook
